Move all classes to duke package
