+++ src/main/java/duke/Duke.java
@@ -1,9 +1,4 @@
-import java.util.List;
-import java.util.ArrayList;
-import java.util.Scanner;
-
-
-
+package duke;
 public class Duke {
 
     private Storage storage;
