+++ src/main/java/duke/ToDoCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class ToDoCommand extends Command {
 
     private String description;
