+++ src/main/java/duke/Ui.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.List;
 import java.util.Scanner;
 public class Ui {
